% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaginR.R
\name{ColorPhenotype}
\alias{ColorPhenotype}
\title{Give the color phenotype of the pearl oyster's}
\usage{
ColorPhenotype(picture)
}
\arguments{
\item{picture}{The picture uploaded by load.image()}
}
\value{
The color phenotype of the pearl oyster's inner shell
}
\description{
Returns the color phenotype of the pearl oyster's inner shell (\emph{Pinctada margaritifera})
}
\examples{
fpath <- system.file('extdata/image.jpg',package='ImaginR')
picture <- load.image(fpath)
ColorPhenotype(picture)
}
