% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaginR.R
\name{Forest_cover}
\alias{Forest_cover}
\title{Give the percent of forest covarage}
\usage{
Forest_cover(x, y)
}
\arguments{
\item{x}{path of the original image}
\item{y}{path of the same image but without the soil (only forest cover is keep)}
}
\value{
Percentage of forest cover (0 to 100%)
}
\description{
Give the percent of forest covarage by comparing two pictures : the first one (x) is the original (aerial) image and the second (y) is the same image but without the soil (only forest cover is keep).
}
\examples{
Forest_cover("A1_DJI_0026_cleaned_empty_before.JPG", "A1_DJI_0026_cleaned_empty.JPG")
# 26.82175
}
