% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaginR.R
\name{hex2rgb}
\alias{hex2rgb}
\title{Give RGB color code from HEX color code}
\usage{
hex2rgb(color.vector)
}
\arguments{
\item{color.vector}{HEX code in a vector}
}
\value{
RGB code color
}
\description{
Give a RGB color code matrix from HEX color code vector.
}
\examples{
# RGB color code of "Alice Blue" color:
hex2rgb("#f0f8ff")
# RGB color code of "Red" color:
hex2rgb("#ff0000")
# RGB color code of "Green" color:
hex2rgb("#008000")
}
