% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaginR.R
\name{OneRow}
\alias{OneRow}
\title{Get phenotype, HEX and HSV color code for one picture}
\usage{
OneRow(picture)
}
\arguments{
\item{picture}{The picture uploaded by load.image()}
}
\value{
The HEX and HSV color code and the color phenotype of the pearl oyster's inner shell for one image in one row
}
\description{
Get results in one row
}
\details{
In header:
\itemize{
 \item{id : the name of your pictures}
 \item{h : the hue of the hsv color code}
 \item{s : the saturation of the hsv color code}
 \item{v : the value of the hsv color code}
 \item{hex : the hexadecimal color code}
 \item{phenotype : returns the color phenotype of the pearl oyster's inner shell (\emph{Pinctada margaritifera})}
}
}
\examples{
fpath <- system.file('extdata/image.jpg',package='ImaginR')
picture <- load.image(fpath)
OneRow(picture)
}
