% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaginR.R
\name{OutPutResult}
\alias{OutPutResult}
\title{Get phenotype, HEX and HSV color code for all pictures}
\usage{
OutPutResult(id)
}
\arguments{
\item{id}{The name of the pictures in your working directory}
}
\value{
The HEX and HSV color code and the color phenotype of the pearl oyster's inner shell for all images in a results.csv file
}
\description{
Get results in a .txt file, .csv file and in R data.frame
This function does what all the others functions do in a very simple way. Just put your images in your working directory (don't forget to getwd() !), do library this package and paste this only code: "OutPutResult()". You will get the results into your consol and in a results.csv file in your working directory.
}
\details{
In results.csv:
\itemize{
 \item{id : the name of your pictures}
 \item{h : the hue of the hsv color code}
 \item{s : the saturation of the hsv color code}
 \item{v : the value of the hsv color code}
 \item{hex : the hexadecimal color code}
 \item{phenotype : returns the color phenotype of the pearl oyster's inner shell (\emph{Pinctada margaritifera})}
}
}
