% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaginR.R
\name{PictureResults}
\alias{PictureResults}
\title{Give HEX and HSV color code}
\usage{
PictureResults(picture)
}
\arguments{
\item{picture}{The picture uploaded by load.image()}
}
\value{
results
the HEX and HSV color code
}
\description{
To get the mean of the HEX color code and the HSV color code of the picture
}
\details{
\itemize{
 \item{h : the hue of the hsv color code}
 \item{s : the saturation of the hsv color code}
 \item{v : the value of the hsv color code}
 \item{hex : the hexadecimal color code}
}
}
\examples{
fpath <- system.file('extdata/image.jpg',package='ImaginR')
picture <- load.image(fpath)
PictureResults(picture)
}
