% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaginR.R
\name{print.myclass}
\alias{print.myclass}
\title{Usefull for others functions}
\usage{
\method{print}{myclass}(x, ...)
}
\arguments{
\item{x}{the results}

\item{...}{if necessary}
}
\value{
the header of the results
}
\description{
Function to print the results (usefull for the others functions)
}
